fixes and improvements for easier plugin use and some renames
I am removing the interface and the component "owner"s. That way if someone wants to add their own functionality for their specific graph nodes, they can inherit from the graph instead of having to modify the interface / component in the plugin. Not to mention, modifying the existing solution to support in-project data types on the C++ side of things will create a circular dependency.